# fx-helper-sca-client/src/main/resources/fx-helper-shop-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://fx-helper-shop"}],"paths":{"/update":{"post":{"tags":["Shop"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestShopUpdateParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}}}}},"/switch":{"post":{"tags":["Shop"],"operationId":"switchFx","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestSwitchParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}}}}},"/internal/get_user_shop_list":{"post":{"tags":["Shop"],"operationId":"internalGetUserShopList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalGetUserShopListRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetShopListResult"}}}}}}},"/internal/get_shop_list":{"post":{"tags":["Shop"],"operationId":"internalGetShopList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalGetShopListRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetShopListResult"}}}}}}},"/internal/get_serial_no":{"post":{"tags":["Shop"],"operationId":"internalGetSerialNo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalGetSerialNoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseInternalGetSerialNoResult"}}}}}}},"/internal/get_detail_info":{"post":{"tags":["Shop"],"operationId":"internalGetDetailInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalGetDetailInfoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetShopDetailInfoResult"}}}}}}},"/get_shop_list":{"post":{"tags":["Shop"],"operationId":"getShopList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetShopListResult"}}}}}}},"/get_qr_codes":{"post":{"tags":["Shop"],"operationId":"getQrCodes","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestGetQrCodesParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetQrCodesResult"}}}}}}},"/get_public_info":{"post":{"tags":["Shop"],"operationId":"getPublicInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestGetPublicShopInfoParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetPublicShopInfoResult"}}}}}}},"/get_detail_info":{"post":{"tags":["Shop"],"operationId":"getDetailInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestGetDetailInfoParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGetShopDetailInfoResult"}}}}}}},"/create":{"post":{"tags":["Shop"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestShopCreateParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseShopCreateResult"}}}}}}}},"components":{"schemas":{"ResponseStatus":{"type":"object","properties":{"errorCode":{"type":"string"},"errorMessage":{"type":"string"},"errors":{"type":"object","additionalProperties":{"type":"string"}}},"required":["errorCode"]},"MobileRequestHead":{"type":"object","properties":{"platform":{"type":"string"},"brand":{"type":"string"},"model":{"type":"string"},"system":{"type":"string"},"deviceId":{"type":"string"},"sign":{"type":"string"},"ip":{"type":"string"},"mac":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"}}},"RequestShopUpdateParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/ShopUpdateParam"}},"required":["head","param"]},"ShopUpdateParam":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"},"address":{"type":"string"},"name":{"type":"string"},"logo":{"type":"string"},"phone":{"type":"string"},"intro":{"type":"string"}},"required":["address","intro","logo","name","phone"]},"BaseResponse":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"}},"required":["status"]},"RequestSwitchParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/SwitchParam"}},"required":["head","param"]},"SwitchParam":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"},"switchValue":{"type":"integer","format":"int32","maximum":1,"minimum":0}},"required":["shopId","switchValue"]},"InternalGetUserShopListRequest":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"}},"required":["userId"]},"GetShopListResult":{"type":"object","properties":{"shopList":{"type":"array","items":{"$ref":"#/components/schemas/ShopDetail"}}}},"ResponseGetShopListResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/GetShopListResult"}},"required":["status"]},"ShopDetail":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"},"ownerUserId":{"type":"integer","format":"int64"},"shopName":{"type":"string"},"logo":{"type":"string"},"intro":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"status":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["address","createdAt","intro","logo","ownerUserId","phone","shopId","shopName","status","updatedAt"]},"InternalGetShopListRequest":{"type":"object","properties":{"shopIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"required":["shopIds"]},"InternalGetSerialNoRequest":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"},"serialNo":{"type":"string"}},"required":["serialNo","shopId"]},"InternalGetSerialNoResult":{"type":"object","properties":{"serialNo":{"type":"string"}},"required":["serialNo"]},"ResponseInternalGetSerialNoResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/InternalGetSerialNoResult"}},"required":["status"]},"InternalGetDetailInfoRequest":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"}},"required":["shopId"]},"GetShopDetailInfoResult":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"},"ownerUserId":{"type":"integer","format":"int64"},"shopName":{"type":"string"},"logo":{"type":"string"},"intro":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"status":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["address","createdAt","intro","logo","ownerUserId","phone","shopId","shopName","status","updatedAt"]},"ResponseGetShopDetailInfoResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/GetShopDetailInfoResult"}},"required":["status"]},"BaseRequest":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"}}},"GetQrCodesParam":{"type":"object","properties":{"number":{"type":"integer","format":"int32","maximum":10},"shopId":{"type":"integer","format":"int64"}},"required":["number","shopId"]},"RequestGetQrCodesParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/GetQrCodesParam"}},"required":["head","param"]},"GetQrCodesResult":{"type":"object","properties":{"qrCodes":{"type":"array","items":{"type":"string"}}}},"ResponseGetQrCodesResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/GetQrCodesResult"}},"required":["status"]},"GetPublicShopInfoParam":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"}},"required":["shopId"]},"RequestGetPublicShopInfoParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/GetPublicShopInfoParam"}},"required":["head","param"]},"GetPublicShopInfoResult":{"type":"object","properties":{"shopName":{"type":"string"},"logo":{"type":"string"},"intro":{"type":"string"},"address":{"type":"string"}},"required":["address","intro","logo","shopName"]},"ResponseGetPublicShopInfoResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/GetPublicShopInfoResult"}},"required":["status"]},"GetDetailInfoParam":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"}},"required":["shopId"]},"RequestGetDetailInfoParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/GetDetailInfoParam"}},"required":["head","param"]},"RequestShopCreateParam":{"type":"object","properties":{"head":{"$ref":"#/components/schemas/MobileRequestHead"},"param":{"$ref":"#/components/schemas/ShopCreateParam"}},"required":["head","param"]},"ShopCreateParam":{"type":"object","properties":{"address":{"type":"string"},"name":{"type":"string"},"logo":{"type":"string"},"phone":{"type":"string"},"intro":{"type":"string"},"paySecret":{"type":"string"}},"required":["address","intro","logo","name","paySecret","phone"]},"ResponseShopCreateResult":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ResponseStatus"},"result":{"$ref":"#/components/schemas/ShopCreateResult"}},"required":["status"]},"ShopCreateResult":{"type":"object","properties":{"shopId":{"type":"integer","format":"int64"}},"required":["shopId"]}}}}