# fx-helper-sca-shop/src/main/java/com/iqexception/fxhelper/shop/controller/dto/InternalGetDetailInfoRequest.java
package com.iqexception.fxhelper.shop.controller.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

public class InternalGetDetailInfoRequest {

    @Positive
    @NotNull
    private Long shopId;

    public Long getShopId() {
        return shopId;
    }

    public void setShopId(Long shopId) {
        this.shopId = shopId;
    }
}