# fx-helper-sca-shop/src/main/java/com/iqexception/fxhelper/shop/controller/dto/InternalGetSerialNoRequest.java
package com.iqexception.fxhelper.shop.controller.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;

public class InternalGetSerialNoRequest {
    @NotNull
    private Long shopId;
    @NotBlank
    private String serialNo;

    public Long getShopId() {
        return shopId;
    }

    public void setShopId(Long shopId) {
        this.shopId = shopId;
    }

    public String getSerialNo() {
        return serialNo;
    }

    public void setSerialNo(String serialNo) {
        this.serialNo = serialNo;
    }
}